Refactor Resource APIs
