+++ r2-shared-swift/Fetcher/Resource/Resource.swift
@@ -117,152 +117,6 @@ public enum ResourceError: Swift.Error {
 /// If the transformation doesn't apply, simply return resource unchanged.
 public typealias ResourceTransformer = (Resource) -> Resource
 
-/// Creates a Resource that will always return the given `error`.
-public final class FailureResource: Resource {
-
-    private let error: ResourceError
-    
-    public init(link: Link, error: ResourceError) {
-        self.link = link
-        self.error = error
-    }
-    
-    public let link: Link
-    
-    public var length: ResourceResult<UInt64> { .failure(error) }
-
-    public func read(range: Range<UInt64>?) -> ResourceResult<Data> {
-        return .failure(error)
-    }
-    
-    public func close() {}
-
-}
-
-/// Creates a `Resource` serving raw data.
-public final class DataResource: Resource {
-    
-    public typealias Factory = () -> (link: Link, data: ResourceResult<Data>)
-
-    private let make: Factory
-    private lazy var result: (link: Link, data: ResourceResult<Data>) = make()
-
-    /// Creates a `Resource` serving an array of bytes.
-    public init(make: @escaping Factory) {
-        self.make = make
-    }
-    
-    public convenience init(link: Link, makeData: @escaping (inout Link) throws -> Data = { _ in Data() }) {
-        self.init {
-            var link = link
-            let data = ResourceResult<Data> { try makeData(&link) }
-            return (link: link, data: data)
-        }
-    }
-    
-    public convenience init(link: Link, data: Data) {
-        self.init {
-            return (link: link, data: .success(data))
-        }
-    }
-    
-    /// Creates a `Resource` serving a string encoded as UTF-8.
-    public convenience init(link: Link, string: String) {
-        // It's safe to force-unwrap when using a unicode encoding.
-        // https://www.objc.io/blog/2018/02/13/string-to-data-and-back/
-        self.init(link: link, data: string.data(using: .utf8)!)
-    }
-    
-    public var link: Link { result.link }
-
-    public var length: ResourceResult<UInt64> { result.data.map { UInt64($0.count) } }
-
-    public func read(range: Range<UInt64>?) -> ResourceResult<Data> {
-        return result.data.map { data in
-            let length = UInt64(data.count)
-            if let range = range?.clamped(to: 0..<length) {
-                return data[range]
-            } else {
-                return data
-            }
-        }
-    }
-    
-    public func close() {}
-    
-}
-
-/// A base class for a `Resource` which acts as a proxy to another one.
-///
-/// Every function is delegating to the proxied resource, and subclasses should override some of
-/// them.
-open class ResourceProxy: Resource {
-
-    public let resource: Resource
-    
-    public init(_ resource: Resource) {
-        self.resource = resource
-    }
-    
-    open var link: Link { resource.link }
-
-    open var length: ResourceResult<UInt64> { resource.length }
-    
-    open func read(range: Range<UInt64>?) -> ResourceResult<Data> {
-        return resource.read(range: range)
-    }
-    
-    open func close() {
-        resource.close()
-    }
-}
-
-/// A `Resource` proxy which applies a transformation on the original content.
-public class TransformingResource: ResourceProxy {
-    
-    private let transform: (Data) -> Data
-    
-    public init(_ resource: Resource, transform: @escaping (Data) -> Data) {
-        self.transform = transform
-        super.init(resource)
-    }
-    
-    public override var length: ResourceResult<UInt64> {
-        transformedData.map { UInt64($0.count) }
-    }
-    
-    public override func read(range: Range<UInt64>?) -> ResourceResult<Data> {
-        return transformedData.map { data in
-            if let range = range?.clamped(to: 0..<UInt64(data.count)) {
-                return data[range]
-            } else {
-                return data
-            }
-        }
-    }
-
-    private lazy var transformedData: ResourceResult<Data> =
-        resource.read().map(transform)
-
-}
-
-/// Convenient shortcuts to create a `TransformingResource`.
-public extension Resource {
-    
-    func map(transform: @escaping (Data) -> Data) -> Resource {
-        return TransformingResource(self, transform: transform)
-    }
-    
-    func mapAsString(encoding: String.Encoding? = nil, transform: @escaping (String) -> String) -> Resource {
-        let encoding = encoding ?? link.mediaType?.encoding ?? .utf8
-        return TransformingResource(self) { data in
-            let string = String(data: data, encoding: encoding) ?? ""
-            return transform(string).data(using: .utf8) ?? Data()
-        }
-    }
-    
-}
-
 public typealias ResourceResult<Success> = Result<Success, ResourceError>
 
 public extension Result where Failure == ResourceError {
